requirements.txt
Python 3.8+
PyTorch 2.0+
OpenCV
NumPy
Matplotlib
Seaborn
Plotly
PIL
elasticdeform